==> notebooks/_deprecated/02-Filter.ipynb <==
import pandas as pd
import os
from constants import EVALUATION_DATE_FOLDER

# Filter

This step mainly remove less relevant resources to narrow down our evaluation.

This notebooks creates a file, i.e., `[RESOURCE_TYPE]_filtered_ids.csv` that contains the `id`s of resources that are selected for evaluation.

# Make sure to set the proper path for saving the filtered data
EVALUATION_DATE_FOLDER

## Data Portals

Nothing to filter out for data portals yet.

PATH_GLOBAL_ID_TABLE = os.path.join('..', 'output', 'data-portal_id_map.csv')
PATH_FILTERED_ID_TABLE = os.path.join('..', 'output', EVALUATION_DATE_FOLDER, 'data-portal_filtered_ids.csv')

df_map = pd.read_csv(PATH_GLOBAL_ID_TABLE)
df_map[['id', 'source_id']].to_csv(PATH_FILTERED_ID_TABLE, index=False)

## Journals

# Somehow, `sjr_` is repeated. This is just a quick fix...
# _ = pd.read_csv(os.path.join('..', 'output', EVALUATION_DATE_FOLDER, 'journal-portal_metadata.csv'))
# _.source_id = _.source_id.apply(lambda x: x.replace('sjr_sjr_', 'sjr_'))
# _.to_csv(os.path.join('..', 'output', EVALUATION_DATE_FOLDER, 'journal-portal_metadata.csv'), index=False)

PATH_FILTERED_ID_TABLE = os.path.join('..', 'output', EVALUATION_DATE_FOLDER, 'journal-portal_filtered_ids.csv')
PATH_METADATA = os.path.join('..', 'output', EVALUATION_DATE_FOLDER, 'journal-portal_metadata.csv')

"""
SJR contains journals that are not related to life sciences. We want to filter them out.
"""
df_meta = pd.read_csv(PATH_METADATA)

# If you want to see the full list of areas, uncomment the below code
# unique_areas = set()
# for item in set(df_meta.areas):
#     for area in item.split(';'):
#         unique_areas.add(area)
# print(unique_areas)

# Interested in the following areas
areas_interested = [
    'Biochemistry, Genetics and Molecular Biology',
    'Health Professions',
    'Immunology and Microbiology',
    'Medicine',
    'Multidisciplinary',
    'Neuroscience',
    'Pharmacology, Toxicology and Pharmaceutics',
    'Psychology',
    'Agricultural and Biological Sciences'
]

# Filter out the metadata
df_meta_filtered = df_meta[df_meta.areas.str.contains('|'.join(areas_interested))]

print(len(df_meta) - len(df_meta_filtered), 'journals are filtered out.', len(df_meta_filtered), 'journals are kept.')

df_meta_filtered[['id', 'source_id']].to_csv(PATH_FILTERED_ID_TABLE, index=False)
